// servicos-global/produto/processo/client/src/shared/lista/pedidoTypes.ts
export type PedidoStatus = string
// ...
export interface Pedido {
  id: string
  tenant_id: string
  company_id: string
  tipo_operacao: TipoOperacao
  numero_pedido: string
  status: PedidoStatus
// ...
  unidade_comercializada_pedido?: string | null
// ...
}